requirements.txt
Flask
Flask-RESful
Flask-SQLAlchemy
requests
json
jwt
Blueprint